% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge_prism.R
\name{model_run}
\alias{model_run}
\title{Run ACCEPT Model}
\usage{
model_run(model_input = NULL)
}
\arguments{
\item{model_input}{A list/json object with "patient_data", "random_sampling_N", "random_distribution_iteration",
and "calculate_CIs" as columns}
}
\value{
Returns a list of results
}
\description{
This function calls the predictACCEPT model. 
If no model_input is specified, it will use the defaults
If one of the columns in model_input is missing, it will replace with the default column
It will check all the columns of the patient_data object; if one is missing, it will
replace with the default, and will remove any extra columns
}
